Testing to see if the drawNode works
This is a long commit message to explain my thought. Read on!

The original idea was to using the drawNode to create each hex node,
and for each hexnode, we attach a event listener to it and listen to
the touch/click event. However, after the test, I found some problems
with this approach.

1. The drawnode does not provide a good way of knowing whether it is
clicked. The cocos2d DrawNode provide a way to create rectangular
boundingbox, and check if the touch coordinate is inside the bounding
box. However, our hex grid cannot fit in this scheme easily. The current
implementation shows when I click a hexnode, many of them can be
triggered.

2. I can get around with the limitation if using DrawNode, instead, I
use Sprite/menuitem/button class to make every hex a
Sprite/menuitem/button. But, I do not feel like to go down this
approach as I kind feel I am not doing it in a right way. For example,
hex is not moving, so it is not a sprite. And it is not a menu, so
using menuitem seems not right. Same for the button, a hex grid of
button doesn't sounds right.

SO, I decide to go down the third path, which is to implement a hex
algerbra calculator to calculate the which hex does the touch point
belongs to. I am not sure this is the best approach, but I think this
is the cleanest way to do this task.

SO, let me implement and see how it goes.